Migrate to aragonOS 4

+++ shared/test-helpers/ganache-cli.sh
@@ -14,10 +14,11 @@ testrpc_running() {
 }
 
 start_testrpc() {
+  echo "Starting our own testrpc instance at port $testrpc_port"
   if [ "$SOLIDITY_COVERAGE" = true ]; then
-    node_modules/.bin/testrpc-sc -i 16 --gasLimit 0xfffffffffff --port "$testrpc_port"  > /dev/null &
+    npx testrpc-sc -i 16 --gasLimit 0xfffffffffff --port "$testrpc_port"  > /dev/null &
   else
-    node_modules/.bin/ganache-cli -i 15 --gasLimit 50000000 --port "$testrpc_port" > /dev/null &
+    npx ganache-cli -i 15 --gasLimit 50000000 --port "$testrpc_port" > /dev/null &
   fi
 
   testrpc_pid=$!
@@ -28,7 +29,6 @@ if testrpc_running; then
   kill -9 $(lsof -i:$testrpc_port -t)
 fi
 
-echo "Starting our own testrpc instance at port $testrpc_port"
 start_testrpc
 sleep 5
 
@@ -36,13 +36,15 @@ sleep 5
 set +e
 result=0
 if [ "$SOLIDITY_COVERAGE" = true ]; then
-  ./node_modules/.bin/solidity-coverage "$@"
+  npx solidity-coverage "$@"
   result=$?
 elif [ "$TRUFFLE_TEST" = true ]; then
   truffle test --network rpc "$@"
   result=$?
 fi
 
-kill -9 $testrpc_pid
+if [ ! -z "$testrpc_pid" ]; then
+  kill -9 $testrpc_pid
+fi
 
 exit $result
